Zero-configuration VisualEditor + PHP for MediaWiki release 1.36
This copies the Parsoid extension code into includes/VEParsoid
to allow a "one extension, zero-configuration" install of
VisualEditor for MW's release. The Parsoid code has been
re-namespaced (`VEParsoid` instead of `MWParsoid`) to avoid
autoloader conflicts if you actually install Parsoid as an
extension (as we do in Wikimedia production). Similarly, we
arrange that the ServiceWiring and RestRoutes configurations
are skipped unless running in zeroconf mode, to avoid
conflicts with the Parsoid extension.

This import matches Parsoid commit
da46f35324478afe319b311044a9d17738a7cb29.

It is not-quite-a-cherry-pick from VE commit
4b04f413fac75e37fbf31029e30fdd0888cb4336.

Bug: T248343
Bug: T261220
Change-Id: Ic63ce40f59c4be8f4fdc5f9ac17798353fc86866