Zero-configuration VisualEditor + PHP for MediaWiki release 1.38
This copies the Parsoid extension code into includes/VEParsoid
to allow a "one extension, zero-configuration" install of
VisualEditor for MW's release. The Parsoid code has been
re-namespaced (`VEParsoid` instead of `MWParsoid`) to avoid
autoloader conflicts if you actually install Parsoid as an
extension (as we do in Wikimedia production). Similarly, we
arrange that the ServiceWiring and RestRoutes configurations
are skipped unless running in zeroconf mode, to avoid
conflicts with the Parsoid extension.

This import matches the Parsoid commit tagged v0.15.0-a26
(9ec4dac79ae3d926507817137d113204d0947958).

It is not-quite-a-cherry-pick from VE commit
75529979de594e37d1e890431d30dce750240178.

Notable: the $wgVisualEditorParsoidSettings configuration option has
been hard deprecated in this release (a warning is issued in the logs)
as it will be replaced by $wgParsoidSettings in core for the next
release.

Bug: T305107
Change-Id: Iac8249bc80dcc0fac3c9cf621e7fb1502e09872e