Always add 1 prefixsearch match when searching for templates
We do this additional prefixsearch[*] anyway. What we did before
was ignoring the result when it was not a 100% exact match.
Instead we can always add this 1 prefixsearch result when it
was not already part of the CirrusSearch result set.

This won't happen often. Usually the 1st prefixsearch result
was already part of the CirrusSearch result set anyway. But
if it wasn't, that's a serious issue for expert users that
expect the search to behave similar to the suggester at the
top of the MediaWiki interface (which is also a prefixsearch).

This backport is cherry-picked from
I959d2b058a3d64596a8cfbe5476ab351e40f8760, but had to be
rewritten because the master branch is already very different.
I tried hard to make this patch as small as possible.

[*] The OpenSearch API mentioned in the code uses the exact
same search backend as prefixsearch.

Change-Id: I959d2b058a3d64596a8cfbe5476ab351e40f8760